+++ src/sorting-visualizer/store/config.ts
@@ -3,12 +3,6 @@ import { algoList } from "../sorting-algorithms/algoList";
 export const menuItems = algoList.map((item) => item.name).concat("all");
 export const initialArray = [6, 8, 3, 5, 1, 9, 2, 7, 4];
 
-export const selectedList = algoList.map((list) => {
-  const{fn, ...rest} = list;
-
-  return{...rest, selected: true}
-});
-
 export const cellCSS = {
   size: 50,
   margin: 4,
@@ -19,6 +13,11 @@ export const colors = {
   highlight: "yellow",
   pivot: "orange",
 };
+
+export const selectedList = algoList.map((list) => {
+  const { fn, ...rest } = list;
+  return { ...rest, selected: true };
+});
 
 const root = document.querySelector(":root") as HTMLElement;
 root.style.setProperty("--cell-size", `${cellCSS.size}px`);
